requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bidict==0.23.1
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cryptography==46.0.5
deprecation==2.1.0
dnspython==2.7.0
eventlet==0.40.3
fastapi==0.134.0
Flask==3.1.2
Flask-SocketIO==5.5.1
fsspec==2026.2.0
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
packaging==25.0
pefile==2023.2.7
postgrest==2.28.0
propcache==0.4.1
psycopg2==2.9.11
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.0
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.6
PyJWT==2.11.0
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.2
python-socketio==5.13.0
pywin32-ctypes==0.2.3
realtime==2.28.0
requests==2.32.5
rich==14.3.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
storage3==2.28.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.0
supabase-auth==2.28.0
supabase-functions==2.28.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zstandard==0.25.0
